+++ src/pages/Store/type/model/StoreModel.ts
@@ -1,0 +1,8 @@
+export type StoreModel = {
+  userId: number;
+  name: string;
+  adress: string;
+  phoneNumber: string;
+  floorSpace: number;
+  seatingCapacity: number;
+};
